BUG: Deprecated classes were still being compiled with VTK_LEGACY_REMOVE
The issue was that CMake's `file(GLOB..)` returns full paths to all
files. Instead of truncating the path later on, this change tells the
GLOB operator to return the file paths relative to the current source
directory. This ensures that the deprecated files get removed by name
only.

Also, fixed a typo in the GeoGraticle.